Добавлено разбиение на классы.
